fix price

+++ api/src/main/java/tpo/uade/api/dto/ProductDto.java
@@ -1,7 +1,6 @@
 package tpo.uade.api.dto;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -13,7 +12,7 @@ import lombok.NoArgsConstructor;
 public class ProductDto implements Serializable {
     private int productId;
     private String name;
-    private BigDecimal price;
+    private double price;
     private String categoryName;
     private String urlImage;
     private String description;
